Gather all syscall use in the osutil helper
Makes it easier to add windows portability
